# swagger.yml
swagger: "2.0"
info:
  title: API de Gestão de Clínicas
  description: API para gerenciamento de clínica.
  version: "1.0.0"
host: localhost:5000
basePath: /
schemes:
  - http
...
paths:
...
  /api/v1/procedimentos:
    get:
      tags:
        - Procedimentos
      summary: Listar procedimentos de forma paginada (protegido)
      parameters:
        - name: pagina
          in: query
          type: integer
          default: 1
        - name: itensPorPagina
          in: query
          type: integer
          default: 2
        - name: Authorization
          in: header
          required: true
          type: string
      responses:
        200:
          description: Lista de procedimentos retornada
          examples:
            application/json:
              pagina: 1
              totalPaginas: 1
              itensPorPagina: 2
              registrosTotais: 2
              procedimentos:
                - id: 1
                  nome: Limpeza Dental
                  desc: Procedimento de limpeza completa dos dentes
                  valorPlano: 100.0
                  valorParticular: 200.0
                - id: 2
                  nome: Restauração
                  desc: Restauração de cárie
                  valorPlano: 150.0
                  valorParticular: 300.0
        401:
          description: Não autorizado
        404:
          description: Página solicitada não disponível
        500:
          description: Erro interno
    post:
      tags:
        - Procedimentos
      summary: Adicionar um novo procedimento (protegido) (admin)
      consumes:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              nome:
                type: string
              desc:
                type: string
              valorPlano:
                type: number
                format: float
              valorParticular:
                type: number
                format: float
          examples:
            application/json:
              nome: Limpeza Dental
              desc: Procedimento de limpeza completa dos dentes
              valorPlano: 100.0
              valorParticular: 200.0
      responses:
        200:
          description: Procedimento criado com sucesso
          examples:
            application/json:
              msg: Procedimento criado com sucesso
              id: 1
        400:
          description: Campos obrigatórios não informados ou valores inválidos
        401:
          description: Não autorizado
        403:
          description: Acesso negado
        409:
          description: Já existe um procedimento com esse nome
        500:
          description: Erro interno
  /api/v1/procedimentos/{id}:
    get:
      tags:
        - Procedimentos
      summary: Obter procedimento por ID (protegido)
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - name: Authorization
          in: header
          required: true
          type: string
      responses:
        200:
          description: Procedimento encontrado
          examples:
            application/json:
              id: 1
              nome: Limpeza Dental
              desc: Procedimento de limpeza completa dos dentes
              valorPlano: 100.0
              valorParticular: 200.0
        401:
          description: Não autorizado
        404:
          description: Procedimento não encontrado
        500:
          description: Erro interno
    put:
      tags:
        - Procedimentos
      summary: Alterar um procedimento (protegido) (admin)
      consumes:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - name: Authorization
          in: header
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              nome:
                type: string
              desc:
                type: string
              valorPlano:
                type: number
                format: float
              valorParticular:
                type: number
                format: float
          examples:
            application/json:
              nome: Limpeza Dental Completa
              desc: Procedimento completo de limpeza e polimento dental
              valorPlano: 120.0
              valorParticular: 250.0
      responses:
        200:
          description: Procedimento alterado com sucesso
          examples:
            application/json:
              msg: Procedimento atualizado com sucesso
        400:
          description: Campos obrigatórios não informados ou valores inválidos
        401:
          description: Não autorizado
        403:
          description: Acesso negado
        404:
          description: Procedimento não encontrado
        409:
          description: Já existe outro procedimento com esse nome
        500:
          description: Erro interno
    delete:
      tags:
        - Procedimentos
      summary: Deletar um procedimento (protegido) (admin)
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - name: Authorization
          in: header
          required: true
          type: string
      responses:
        200:
          description: Procedimento deletado com sucesso
          examples:
            application/json:
              msg: Procedimento removido com sucesso
        401:
          description: Não autorizado
        403:
          description: Acesso negado
        404:
          description: Procedimento não encontrado
        409:
          description: Não é possível excluir - procedimento está em uso em atendimentos
        500:
          description: Erro interno
...